More defensive cache get/set
We encountered an issue where, during cache set, the code was trying to
call `decode` on a `str` and erroring. This fix seems to address the
issue, though the check is a bit blunt and there may be an issue
elsewhere that is causing the wrong type of object to be passed around.